feat: 공연자 리뷰 작성 api 구현 #41

+++ src/main/java/umc/ShowHoo/web/spaceReview/entity/SpaceReview.java
@@ -19,7 +19,7 @@ public class SpaceReview {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private int grade; // 리뷰 평점
+    private double grade; // 리뷰 평점
     private String content; // 리뷰 내용
 
     @ManyToOne
